+++ tests/functional/test_subrack.py
@@ -18,6 +18,7 @@ from __future__ import annotations
 import enum
 import json
 import time
+from typing import Generator
 
 import pytest
 import tango
@@ -63,13 +64,43 @@ def subrack_device_fixture(
     return functional_test_context.get_subrack_device(subrack_id)
 
 
+@pytest.fixture(name="fan_mode_cleanup")
+def fan_mode_cleanup_fixture(
+    subrack_device: tango.DeviceProxy,
+) -> Generator:
+    """
+    Cleanup fixture to set fan modes and speeds to max after test.
+
+    :param subrack_device: the subrack Tango device under test.
+
+    :yields: control back to the test
+    """
+    yield
+    # Make sure the subracks are in MANUAL at 100%.
+    # Fans are set in pairs: 1+2, 3+4.
+    print("Setting fans to MANUAL and 100% during teardown.")
+    fan_modes = list(subrack_device.subrackFanModes)
+    if FanMode.AUTO in fan_modes:
+        for fan in [1, 3]:
+            encoded_arg = json.dumps({"fan_id": fan, "mode": int(FanMode.MANUAL)})
+            subrack_device.SetSubrackFanMode(encoded_arg)
+
+    for fan in [1, 3]:
+        encoded_arg = json.dumps({"subrack_fan_id": fan, "speed_percent": 100})
+        subrack_device.SetSubrackFanSpeed(encoded_arg)
+
+
 @scenario("features/subrack.feature", "Monitor and control subrack fan speed")
-def test_monitor_and_control_subrack_fan_speed() -> None:
+def test_monitor_and_control_subrack_fan_speed(
+    fan_mode_cleanup: Generator,
+) -> None:
     """
     Run a test scenario that monitors and controls a subrack fan's speed.
 
     Any code in this scenario function is run at the *end* of the
     scenario.
+
+    :param fan_mode_cleanup: Fixture to ensure fan cleanup after test.
     """
 
 
@@ -129,9 +160,9 @@ def check_subrack_is_online_and_on(
         )
         print("Subrack device is in UNKNOWN state.")
 
-    change_event_callbacks["subrack_state"].assert_change_event(
-        OneOf(tango.DevState.OFF, tango.DevState.ON), lookahead=6
-    )
+        change_event_callbacks["subrack_state"].assert_change_event(
+            OneOf(tango.DevState.OFF, tango.DevState.ON), lookahead=6
+        )
     state = subrack_device.state()
 
     if state == tango.DevState.OFF:
@@ -148,14 +179,14 @@ def check_subrack_is_online_and_on(
     print("Subrack device is in ON state.")
 
 
-@given("a choice of subrack fan", target_fixture="fan_number")
-def choose_a_fan() -> int:
+@given("two subrack fan pairs", target_fixture="fan_number")
+def choose_a_fan() -> list[int]:
     """
-    Return a fan number.
+    Return fan numbers.
 
-    :return: a fan number.
+    :return: fan numbers.
     """
-    return 1
+    return [1, 3]
 
 
 @given("a choice of TPM", target_fixture="tpm_number")
@@ -193,18 +224,17 @@ def choose_a_tpm(
     return 1 + tpms_present.index(True)
 
 
-@given("the fan mode is manual")
+@given("all fan modes are manual")
 def ensure_subrack_fan_mode(
     subrack_device: tango.DeviceProxy,
-    fan_number: int,
+    fan_number: list[int],
     change_event_callbacks: MockTangoEventCallbackGroup,
 ) -> None:
     """
-    Ensure that the fan is in manual mode.
+    Ensure that the fans are in manual mode.
 
     :param subrack_device: the subrack Tango device under test.
-    :param fan_number: number of the subrack fan being exercised by this
-        test.
+    :param fan_number: Fan numbers to target in tests.
     :param change_event_callbacks: dictionary of Tango change event
         callbacks with asynchrony support.
     """
@@ -239,29 +269,32 @@ def ensure_subrack_fan_mode(
     assert fan_modes
 
     expected_fan_modes = fan_modes
-    if expected_fan_modes[fan_number - 1] == FanMode.AUTO:
-        expected_fan_modes[fan_number - 1] = int(FanMode.MANUAL)
+    for fan in fan_number:  # Subrack fans change mode in pairs (1+2), (3+4).
+        if FanMode.AUTO in [expected_fan_modes[fan - 1], expected_fan_modes[fan]]:
+            expected_fan_modes[fan - 1] = int(FanMode.MANUAL)
+            expected_fan_modes[fan] = int(FanMode.MANUAL)
 
-        encoded_arg = json.dumps({"fan_id": fan_number, "mode": int(FanMode.MANUAL)})
-        subrack_device.SetSubrackFanMode(encoded_arg)
+            encoded_arg = json.dumps({"fan_id": fan, "mode": int(FanMode.MANUAL)})
+            print(f"Setting fans {fan}+{fan+1} mode to MANUAL...")
+            subrack_device.SetSubrackFanMode(encoded_arg)
 
-        change_event_callbacks.assert_change_event(
-            "subrack_fan_mode", expected_fan_modes, lookahead=4
-        )
+            change_event_callbacks.assert_change_event(
+                "subrack_fan_mode", expected_fan_modes, lookahead=4
+            )
     subrack_device.unsubscribe_event(sub_id)
 
 
-@given("the fan's speed setting is 90%")
+@given("all fan speed settings are 90%")
 def ensure_subrack_fan_speed_percent(
     subrack_device: tango.DeviceProxy,
-    fan_number: int,
+    fan_number: list[int],
     change_event_callbacks: MockTangoEventCallbackGroup,
 ) -> None:
     """
     Ensure that the fan is set to 90% speed.
 
     :param subrack_device: the subrack Tango device under test.
-    :param fan_number: number of the subrack fan being exercised by this
+    :param fan_number: list of the subrack fans being exercised by this
         test.
     :param change_event_callbacks: dictionary of Tango change event
         callbacks with asynchrony support.
@@ -275,55 +308,63 @@ def ensure_subrack_fan_speed_percent(
         change_event_callbacks["subrack_fan_speeds_percent"],
     )
     change_event_callbacks.assert_change_event(
-        "subrack_fan_speeds_percent", expected_fan_speeds_percent, lookahead=4
+        "subrack_fan_speeds_percent",
+        expected_fan_speeds_percent,
+        lookahead=4,
     )
 
-    # TODO: There is a server-side bug in handling of SetSubrackFanSpeed.
-    # All we see is a HTTP timeout.
-    # And the fan speed setting is never updated.
-    # This scenario cannot be developed further until this bug is fixed.
-    pytest.xfail(reason="Server-side bug")
-
-    speed_percent = fan_speeds_percent[fan_number - 1]
-    if speed_percent != pytest.approx(90.0):
-        encoded_arg = json.dumps({"subrack_fan_id": fan_number, "speed_percent": 90.0})
-        subrack_device.SetSubrackFanSpeed(encoded_arg)
-        expected_fan_speeds_percent[fan_number - 1] = pytest.approx(90.0)
-        change_event_callbacks.assert_change_event(
-            "subrack_fan_speeds_percent", expected_fan_speeds_percent, lookahead=2
-        )
+    for fan in fan_number:
+        speed_percent = fan_speeds_percent[fan - 1]
+        if speed_percent != pytest.approx(90.0):
+            encoded_arg = json.dumps({"subrack_fan_id": fan, "speed_percent": 90.0})
+            subrack_device.SetSubrackFanSpeed(encoded_arg)
+            # Subrack fans update in pairs.
+            expected_fan_speeds_percent[fan - 1] = pytest.approx(90.0)
+            expected_fan_speeds_percent[fan] = pytest.approx(90.0)
+            change_event_callbacks.assert_change_event(
+                "subrack_fan_speeds_percent", expected_fan_speeds_percent, lookahead=2
+            )
 
 
 @given("the fan's speed is approximately 90% of its maximum")
 def ensure_subrack_fan_speed(
     subrack_device: tango.DeviceProxy,
-    fan_number: int,
+    fan_number: list[int],
     change_event_callbacks: MockTangoEventCallbackGroup,
+    station_label: str,
 ) -> None:
     """
     Ensure that the fan's speed is about 90% of its maximum.
 
     :param subrack_device: the subrack Tango device under test.
-    :param fan_number: number of the subrack fan being exercised by this
+    :param fan_number: list of the subrack fans being exercised by this
         test.
     :param change_event_callbacks: dictionary of Tango change event
         callbacks with asynchrony support.
+    :param station_label: The station under test.
     """
-    fan_speeds_percent = list(subrack_device.subrackFanSpeedsPercent)
-    assert fan_speeds_percent[fan_number - 1] == pytest.approx(90.0)  # just checkin'
+    # The subrack at RAL appears to have different maximum fan speeds
+    # for each fan present. Fan2 also consistently reports 0rpm.
+    # Until this issue is clarified we skip checks on fan speed in RPM.
+    if station_label != "stfc-ral-2":
+        fan_speeds_percent = list(subrack_device.subrackFanSpeedsPercent)
+        for fan in fan_number:
+            assert fan_speeds_percent[fan - 1] == pytest.approx(90.0)  # just checkin'
+            assert fan_speeds_percent[fan] == pytest.approx(90.0)  # just checkin'
 
-    expected_fan_speeds = [
-        pytest.approx(p * MAX_SUBRACK_FAN_SPEED / 100.0) for p in fan_speeds_percent
-    ]
+        expected_fan_speeds = [
+            pytest.approx(p * MAX_SUBRACK_FAN_SPEED / 100.0, abs=10)
+            for p in fan_speeds_percent
+        ]
 
-    subrack_device.subscribe_event(
-        "subrackFanSpeeds",
-        tango.EventType.CHANGE_EVENT,
-        change_event_callbacks["subrack_fan_speeds"],
-    )
-    change_event_callbacks.assert_change_event(
-        "subrack_fan_speeds", expected_fan_speeds, lookahead=4
-    )
+        subrack_device.subscribe_event(
+            "subrackFanSpeeds",
+            tango.EventType.CHANGE_EVENT,
+            change_event_callbacks["subrack_fan_speeds"],
+        )
+        change_event_callbacks.assert_change_event(
+            "subrack_fan_speeds", expected_fan_speeds, lookahead=10
+        )
 
 
 @given(parsers.parse("the TPM is {target_power}"))
@@ -382,17 +423,18 @@ def ensure_tpm_power_state(
 @when("I set the fan speed to 100%")
 def set_subrack_fan_speed(
     subrack_device: tango.DeviceProxy,
-    fan_number: int,
+    fan_number: list[int],
 ) -> None:
     """
     Set the subrack's fan speed to 100%.
 
     :param subrack_device: the subrack Tango device under test.
-    :param fan_number: number of the subrack fan being exercised by this
+    :param fan_number: list of the subrack fans being exercised by this
         test.
     """
-    encoded_arg = json.dumps({"subrack_fan_id": fan_number, "speed_percent": 100.0})
-    subrack_device.SetSubrackFanSpeed(encoded_arg)
+    for fan in fan_number:
+        encoded_arg = json.dumps({"subrack_fan_id": fan, "speed_percent": 100.0})
+        subrack_device.SetSubrackFanSpeed(encoded_arg)
 
 
 @when("I tell the subrack to turn on the TPM")
@@ -426,51 +468,62 @@ def turn_off_all_tpms(
 @then("the fan's speed setting becomes 100%")
 def check_subrack_fan_speed_setting(
     subrack_device: tango.DeviceProxy,
-    fan_number: int,
+    fan_number: list[int],
     change_event_callbacks: MockTangoEventCallbackGroup,
 ) -> None:
     """
     Check that the fan's speed setting becomes 100%.
 
     :param subrack_device: the subrack Tango device under test.
-    :param fan_number: number of the subrack fan being exercised by this
+    :param fan_number: list of the subrack fans being exercised by this
         test.
     :param change_event_callbacks: dictionary of Tango change event
         callbacks with asynchrony support.
     """
     fan_speeds_percent = list(subrack_device.subrackFanSpeedsPercent)
     expected_fan_speeds_percent = [pytest.approx(p) for p in fan_speeds_percent]
-    expected_fan_speeds_percent[fan_number - 1] = pytest.approx(100.0)
+    for fan in fan_number:
+        expected_fan_speeds_percent[fan - 1] = pytest.approx(100.0)
+        expected_fan_speeds_percent[fan] = pytest.approx(100.0)
 
     change_event_callbacks["subrack_fan_speeds_percent"].assert_change_event(
-        expected_fan_speeds_percent
+        expected_fan_speeds_percent, lookahead=5
     )
 
 
 @then("the fan's speed becomes approximately 100% of its maximum")
 def check_subrack_fan_speed(
     subrack_device: tango.DeviceProxy,
-    fan_number: int,
+    fan_number: list[int],
     change_event_callbacks: MockTangoEventCallbackGroup,
+    station_label: str,
 ) -> None:
     """
     Check that the fan's speed becomes approximately 100% of its maximum.
 
     :param subrack_device: the subrack Tango device under test.
-    :param fan_number: number of the subrack fan being exercised by this
+    :param fan_number: list of the subrack fans being exercised by this
         test.
     :param change_event_callbacks: dictionary of Tango change event
         callbacks with asynchrony support.
+    :param station_label: The station under test.
     """
-    fan_speeds_percent = list(subrack_device.subrackFanSpeedsPercent)
-    assert fan_speeds_percent[fan_number - 1] == pytest.approx(100.0)  # just checkin'
+    # The subrack at RAL appears to have different maximum fan speeds
+    # for each fan present. Fan2 also consistently reports 0rpm.
+    # Until this issue is clarified we skip checks on fan speed in RPM.
+    if station_label != "stfc-ral-2":
+        fan_speeds_percent = list(subrack_device.subrackFanSpeedsPercent)
+        for fan in fan_number:
+            assert fan_speeds_percent[fan - 1] == pytest.approx(100.0)  # just checkin'
+            assert fan_speeds_percent[fan] == pytest.approx(100.0)  # just checkin'
 
-    expected_fan_speeds = [
-        pytest.approx(p * MAX_SUBRACK_FAN_SPEED / 100.0) for p in fan_speeds_percent
-    ]
-    change_event_callbacks.assert_change_event(
-        "subrack_fan_speeds", expected_fan_speeds
-    )
+        expected_fan_speeds = [
+            pytest.approx(p * MAX_SUBRACK_FAN_SPEED / 100.0, abs=10)
+            for p in fan_speeds_percent
+        ]
+        change_event_callbacks.assert_change_event(
+            "subrack_fan_speeds", expected_fan_speeds, lookahead=10
+        )
 
 
 @then(parsers.parse("the subrack reports that the TPM is {target_power}"))