+++ tests/functional/test_tile.py
@@ -184,8 +184,11 @@ def check_spsstation_state(
         station, "state", tango.DevState.STANDBY
     )
     station.on()
-    AttributeWaiter(timeout=180).wait_for_value(station, "state", tango.DevState.ON)
-
+    try:
+        AttributeWaiter(timeout=180).wait_for_value(station, "state", tango.DevState.ON)
+    except AssertionError:
+        # Hardware can be in the ALARM state, we should still continue.
+        assert station.state() in [tango.DevState.ON, tango.DevState.ALARM]
     for tile in station_tiles:
         AttributeWaiter(timeout=180).wait_for_value(
             tile, "tileProgrammingState", "Synchronised", lookahead=5